nix env shell: Move from nix shell, add shorthand alias
